Switch temporary directory to use task UUID
Prior to this commit (and if the `use_task_uuid` parameter is
not passed, or is set to `False`), a timestamp and counter
would be used to generate the remote temporary directory.

This guaranteed uniqueness, but at the cost of predictability and
trackability; callback plugins in particular have no way of *knowing*
where the temporary directory is.

Fortunately, there is another method of generating the temporary directory,
which still prevents collisions but also has the property that it can be
derived by anything that has access to the task structure - the task uuid.

**NOTE** I'm not sure if this change should be on some kind of switch.