Add ANSIBLE_TEE_REMOTE_OUTPUT (and ANSIBLE_TMPDIR)
The `ANSIBLE_TEE_REMOTE_OUTPUT` option functions as if the stdout
and stderr streams were tee'd to the temporary directory generated
by `_make_tmp_path`, under the file names `stdout` and `stderr`.

This is particularly handy for use cases where the user wants to
check `stdout` and `stderr` in the middle of a long-running or 'stuck'
process.  Additionally, the previous commit, which set the tmp path
to use the task uuid, means that callback plugins can trivially and
deterministically identify the location of the live streams on the
remote systems.

The `ANSIBLE_TMPDIR` environment variable was added to identify the
base path to use.  When using the ANSIBALLZ wrapper, the task is copied
over as a zip file, extracted to an even-more-temporary temp directory,
and executed from there, with no record of where it was executed from.
Instead of trying to divine the original path, or passing arguments to
the generated script, this sets an environment variable to keep a
permanent record of what Ansible's temporary directory is.  Passing flags
may be more appropriate, but would involve touching quite a few more pieces
of code to follow the paths through, and an `ANSIBLE_TMPDIR` may find more
use anyway (and has a natural parallel in `TMPDIR`).